Fix copy & paste of table.
